Turn GitClient into a class
