+++ AuroraEditor/Base/NavigatorSidebar/UI/SourceControlNavigator/Changes/CommitChangesView.swift
@@ -23,12 +23,7 @@ struct CommitChangesView: View {
 
     init(workspace: WorkspaceDocument) {
         self.workspace = workspace
-        if let folderURL = workspace.workspaceClient?.folderURL {
-            self.gitClient = GitClient.default(
-                directoryURL: folderURL,
-                shellClient: sharedShellClient.shellClient
-            )
-        }
+        self.gitClient = workspace.workspaceClient?.model?.gitClient
     }
 
     var body: some View {
